Gen Ai/pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community",
    "langchain-google-genai",
    "langchain-core>=0.3.69",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "langserve>=0.3.1",
    "sse-starlette>=2.4.1",
]
